requirements.txt
streamlit
opencv-python
ultralytics
openai-whisper
langchain
langchain-google-genai
chromadb
networkx
moviepy
python-dotenv
scikit-learn
matplotlib
numpy>=1.26.0